refactor: Checklist items
Refactored checklist items to have only one checklist notes modal, assigned the mapping to be outside of the now deprecated checklist form item
