+++ src/components/Checklist/Checklist.js
@@ -1,9 +1,14 @@
+import { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { Pane } from '@folio/stripes/components';
 import { useMutation, useQueryClient } from 'react-query';
 import { AppIcon, useNamespace, useOkapiKy } from '@folio/stripes-core';
 import { FormattedMessage } from 'react-intl';
-import ChecklistForm from './ChecklistForm';
+import isEqual from 'lodash/isEqual';
+import differenceWith from 'lodash/differenceWith';
+import ChecklistItem from './ChecklistItem';
+import ChecklistNotesModal from './ChecklistNotesModal';
+import useChecklistItemDefinitions from '../../hooks/useChecklistItemDefinitions';
 import { PUBLICATION_REQUEST_ENDPOINT } from '../../constants/endpoints';
 
 const propTypes = {
@@ -12,17 +17,48 @@ const propTypes = {
 };
 
 const Checklist = ({ onToggle, resource }) => {
+  const itemDefinitions = useChecklistItemDefinitions();
   const [namespace] = useNamespace();
   const queryClient = useQueryClient();
   const ky = useOkapiKy();
 
+  const [selectedNotesItem, setSelectedNotesItem] = useState(false);
+  const [checklistItems, setChecklistItems] = useState([]);
+
+  useEffect(() => {
+    const itemList = itemDefinitions.map((definition) => ({ definition }));
+    const output = [];
+    itemList.forEach((item) => {
+      const relevantItem = resource.checklist.find(
+        (ci) => ci.definition.name === item.definition.name
+      );
+      if (relevantItem) {
+        output.push(relevantItem);
+      } else {
+        output.push(item);
+      }
+    });
+    if (
+      differenceWith(checklistItems, output, isEqual)?.length === 0 &&
+      !isEqual(checklistItems, output)
+    ) {
+      setChecklistItems(output);
+    }
+  }, [resource, itemDefinitions, checklistItems]);
+
   const { mutateAsync: putChecklist } = useMutation(
     ['Checklist', 'putChecklist'],
     (data) => {
-      ky.put(PUBLICATION_REQUEST_ENDPOINT(resource.id), { json: data })
-        .then(() => {
-          queryClient.invalidateQueries([namespace, 'data', 'view', resource?.id]);
-        });
+      ky.put(PUBLICATION_REQUEST_ENDPOINT(resource.id), { json: data }).then(
+        () => {
+          queryClient.invalidateQueries([
+            namespace,
+            'data',
+            'view',
+            resource?.id,
+          ]);
+        }
+      );
     }
   );
 
@@ -39,10 +75,19 @@ const Checklist = ({ onToggle, resource }) => {
       onClose={onToggle}
       paneTitle={<FormattedMessage id="ui-oa.checklist" />}
     >
-      <ChecklistForm
-        checklist={resource.checklist}
-        handleSubmit={handleSubmit}
+      {checklistItems.map((item) => {
+        return (
+          <ChecklistItem
+            handleSubmit={handleSubmit}
+            item={item}
+            setSelectedNotesItem={setSelectedNotesItem}
+          />
+        );
+      })}
+      <ChecklistNotesModal
+        item={selectedNotesItem}
         ownerId={resource.id}
+        setSelectedNotesItem={setSelectedNotesItem}
       />
     </Pane>
   );
